Make EC phone optional and cellphone required for pre-reg Fixes https://github.com/Anthrocon-Reg/anthrocon_plugin/issues/22.
